TASK_ONE/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.2
drf-yasg==1.21.7
frozenlist==1.4.1
geoip2==4.8.0
idna==3.7
inflection==0.5.1
maxminddb==2.6.2
multidict==6.0.5
packaging==24.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.2.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
